[tuner] Move constraint generation out of canddiate_gen. NFC.
This is just code motion to make the code more modular.
